Works on normalization and mean centered on songs
On the tutor song, the fact to normalized and mean centered doesn't change the features values.
The cmp_init_song_model_and_synth_score files show that the current measure of error is not satisfying.
The alpha learned by the model is in fact a proxy of the air pressure in the synthesizer code which is in fact the envelop of the tutor song (see song_fitter_analyser).