New version of measure + night algo optimisation
The new measure use the features from the tutor song to rescale the song features. It implies potential big changes in the code, even though there is not that much currently.
Optimisation the night algorithm: only one call to bsa_measure(tutor_song). The result is now given when needed instead of being recalulated.